+++ src/main/kotlin/racingcar/Application.kt
@@ -1,10 +1,5 @@
 package racingcar
 
-import racingcar.domain.RacingCarNumberGenerator
-import racingcar.domain.RacingGame
-import racingcar.view.InputView
-import racingcar.view.ResultView
-
 fun main() {
     val cars = InputView.readCars()
     val roundCount = InputView.readRoundCount()
