+++ src/test/kotlin/racingcar/CarTest.kt
@@ -2,7 +2,6 @@ package racingcar
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import racingcar.domain.Car
 
 class CarTest : StringSpec({
