+++ src/test/kotlin/racingcar/MovingJudgeRuleTest.kt
@@ -2,7 +2,6 @@ package racingcar
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
-import racingcar.domain.MovingJudgeRule
 
 class MovingJudgeRuleTest : BehaviorSpec({
